Minor bugfixes in python3 port

+++ cube/models/vocoder.py
@@ -36,7 +36,7 @@ class Vocoder:
         self.trainer.set_clip_threshold(5.0)
         # self.trainer = dy.AdamTrainer(self.model)
         # MGCs are extracted at 12.5 ms
-        from utils import orthonormal_VanillaLSTMBuilder
+        from models.utils import orthonormal_VanillaLSTMBuilder
         lstm_builder = orthonormal_VanillaLSTMBuilder
         if runtime:
             lstm_builder = dy.VanillaLSTMBuilder
@@ -80,7 +80,7 @@ class Vocoder:
         self.softmax_fine_b = self.model.add_parameters((256))
 
     def _upsample(self, mgc, start, stop):
-        mgc_index = start / len(self.upsample_w_t)
+        mgc_index = int(start / len(self.upsample_w_t))
         ups_index = start % len(self.upsample_w_t)
         upsampled = []
         mgc_index_next = mgc_index + 1
@@ -147,7 +147,7 @@ class Vocoder:
     def synthesize(self, mgc, batch_size, sample=True, temperature=1.0):
         synth = []
         total_audio_len = mgc.shape[0] * len(self.upsample_w_t)
-        num_batches = total_audio_len / batch_size
+        num_batches = int(total_audio_len / batch_size)
         if total_audio_len % batch_size != 0:
             num_batches + 1
         last_rnn_coarse_state = None
@@ -177,7 +177,7 @@ class Vocoder:
             for index in range(stop - start):
                 w_index += 1
 
-                curr_proc = w_index * 100 / total_audio_len
+                curr_proc = int(w_index * 100 / total_audio_len)
                 if curr_proc % 5 == 0 and curr_proc != last_proc:
                     last_proc = curr_proc
                     sys.stdout.write(' ' + str(curr_proc))
@@ -274,7 +274,7 @@ class Vocoder:
             for index in range(stop - start):
                 w_index += 1
 
-                curr_proc = w_index * 100 / len(wave)
+                curr_proc = int(w_index * 100 / len(wave))
                 if curr_proc % 5 == 0 and curr_proc != last_proc:
                     last_proc = curr_proc
                     sys.stdout.write(' ' + str(curr_proc))
@@ -370,7 +370,7 @@ class VocoderOld:
             self.upsample_b_t.append(self.model.add_parameters((self.UPSAMPLE_PROJ)))
 
         self.output_lookup = self.model.add_lookup_parameters((256, self.OUTPUT_EMB_SIZE))
-        from utils import orthonormal_VanillaLSTMBuilder
+        from models.utils import orthonormal_VanillaLSTMBuilder
         # self.rnn = orthonormal_VanillaLSTMBuilder(self.RNN_LAYERS, self.OUTPUT_EMB_SIZE + self.UPSAMPLE_PROJ, self.RNN_SIZE, self.model)
         self.rnn = dy.VanillaLSTMBuilder(self.RNN_LAYERS, self.OUTPUT_EMB_SIZE + self.UPSAMPLE_PROJ,
                                          self.RNN_SIZE, self.model)
